requirements.txt
certifi==2017.4.17
chardet==3.0.4
dj-database-url==0.3.0
Django==1.8.4
django-configurations==2.0
django-crispy-forms==1.5.1
django-dotenv==1.3.0
django-extensions==1.5.5
djangorestframework==3.6.3
idna==2.5
psycopg2==2.7.1
pytz==2017.2
requests==2.18.1
six==1.9.0
urllib3==1.21.1
waitress==0.8.9
websockets==2.6
whitenoise==2.0.3
